requirements.txt
# Web service dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# Auto dialer project dependencies (required since we import from it)
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
et_xmlfile==2.0.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
gspread==6.2.1
httplib2==0.31.0
idna==3.11
numpy==2.0.2
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.3
proto-plus==1.27.0
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
xlsxwriter==3.2.9

# Database dependencies for settings backend
sqlalchemy>=2.0.0
pymysql>=1.1.0
